Qemu/RiscvVirt: Enable booting EDK2 in S-mode as payload of opensbi
Currently, opensbi is integrated as a library in EDK2. Hence,
it is always required to boot EDK2 in M-mode. This patch enables
booting EDK2 in S-mode. A prior stage M-mode firmware like
opensbi is mandatory in this case to boot edk2 in S-Mode.

This enables working with latest opensbi without going
through the process of integration with EDK2.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>